stfsender: refactor output fairmq backend
